Version bump kotlin, build tools. Make comments view scroll more smoothly

+++ app/src/main/kotlin/se/ntlv/newsbringer/NewsThreadListAdapter.kt
@@ -50,7 +50,7 @@ open class NewsThreadListAdapter(ctx: Context, layout: Int, cursor: Cursor?, fla
         fun TextView.getContent(): String = this.getText().toString()
 
         val metadata: Metadata
-            get() = Metadata(id, text ?: "", title.getContent(), by.getContent(), time.getContent(), score.getContent(), link, commentQuantity ?: 0)
+            get() = Metadata(id as Long, text ?: "", title.getContent(), by.getContent(), time.getContent(), score.getContent(), link, commentQuantity ?: 0)
     }
 
     val View.tag: ViewHolder
